group by player id

+++ src/main/scala/classes/EventHolder.scala
@@ -3,6 +3,7 @@ package classes
 
 import org.apache.spark.sql.SparkSession
 
+
 class EventHolder(spark: SparkSession, var eventID: Long = 0) extends Holder(spark){
     override protected var baseURL: String = s"https://raw.githubusercontent.com/statsbomb/open-data/master/data/events/$eventID.json"
     if(eventID!=0){
@@ -12,9 +13,11 @@ class EventHolder(spark: SparkSession, var eventID: Long = 0) extends Holder(spa
     private def setBaseURL(): Unit =
         baseURL = s"https://raw.githubusercontent.com/statsbomb/open-data/master/data/events/$eventID.json"
 
-    def setEventID(eventID: Int): Unit ={
+
+    def setEventID(eventID: Long): Unit ={
         this.eventID = eventID
         setBaseURL()
         initDF()
     }
+
 }
